Got rid of unnecessary network_security_config.xml and unnecessary activities. Made menu drop down more concise.
