+++ app/src/main/AndroidManifest.xml
@@ -9,9 +9,6 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".AboutUS"></activity>
-        <activity android:name=".GuideLines" />
-        <activity android:name=".InviteUs" />
         <activity android:name=".GetInvolved" />
         <activity android:name=".MainActivity">
             <intent-filter>
