+++ app/src/main/java/com/thinqtv/thinqtv_android/GetInvolved.java
@@ -22,7 +22,6 @@ public class GetInvolved extends AppCompatActivity implements AdapterView.OnItem
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_get_involve);
-
         view = findViewById(R.id.webView);
         view.setLayerType(WebView.LAYER_TYPE_SOFTWARE, null);
         view.getSettings().setJavaScriptEnabled(true);
@@ -41,9 +40,7 @@ public class GetInvolved extends AppCompatActivity implements AdapterView.OnItem
                 }
             }
         });
-
         view.loadUrl("https://thinqtv.herokuapp.com/getinvolved");
-
         Spinner spinner = findViewById(R.id.spinner1);
         ArrayAdapter<CharSequence> adapter = ArrayAdapter.createFromResource(this,R.array.Pages,android.R.layout.simple_spinner_item);
         adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
@@ -51,7 +48,27 @@ public class GetInvolved extends AppCompatActivity implements AdapterView.OnItem
         spinner.setOnItemSelectedListener(this);
     }
 
-
+    public void changeSite(String site){
+        view = findViewById(R.id.webView);
+        view.setLayerType(WebView.LAYER_TYPE_SOFTWARE, null);
+        view.getSettings().setJavaScriptEnabled(true);
+        view.getSettings().setDomStorageEnabled(true);
+        view.setWebViewClient(new WebViewClient() {
+            @Override
+            public void onLoadResource(WebView webView, String url) {
+                try {
+                    // remove the nav bar and the footer from each loaded page.
+                    view.loadUrl("javascript:(window.onload = function() { " +
+                            "(navBar = document.getElementsByTagName('nav')[0]); navBar.parentNode.removeChild(navBar);" +
+                            "(footer = document.getElementsByTagName('footer')[0]); footer.parentNode.removeChild(footer);" +
+                            "})()");
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
+        });
+        view.loadUrl(site);
+    }
 
     public void goHome(View v){
         Intent i = new Intent(this, MainActivity.class);
@@ -59,31 +76,25 @@ public class GetInvolved extends AppCompatActivity implements AdapterView.OnItem
         System.out.println(" ''" + v + " ''");
     }
 
-
-
     @Override
     public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
         switch(position){
             case 0:
+                changeSite("https://thinqtv.herokuapp.com/getinvolved");
                 break;
             case 1:
-                System.out.println("Selected");
-
+                changeSite("https://thinqtv.herokuapp.com/drschaeferspeaking");
                 break;
             case 2:
-                startActivity(new Intent(this,InviteUs.class));
+                changeSite("https://thinqtv.herokuapp.com/bystanderguidelines");
                 break;
             case 3:
-                startActivity(new Intent(this,GuideLines.class));
-                break;
-            case 4:
-                startActivity(new Intent(this,AboutUS.class));
+                changeSite("https://thinqtv.herokuapp.com/aboutus");
                 break;
         }
     }
 
     @Override
     public void onNothingSelected(AdapterView<?> parent) {
-
     }
 }
