--- app/src/main/res/layout/activity_invite_us.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-
-
-    android:background="#ffffff"
-    tools:context=".InviteUs" >
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="#F8F8F8"
-        android:orientation="vertical"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent">
-
-        <Spinner
-            android:id="@+id/spinner1"
-            android:layout_width="50dp"
-            android:layout_height="wrap_content"
-            android:background="@android:drawable/btn_dropdown"
-            android:spinnerMode="dropdown"/>
-
-
-        <ImageButton
-            android:id="@+id/homeButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="0dp"
-            android:adjustViewBounds="true"
-            android:background="@android:color/transparent"
-            android:clickable="false"
-            android:onClick="goHome"
-            app:srcCompat="@mipmap/ic_launcher_foreground" />
-
-        <WebView
-            android:id="@+id/webView"
-            android:layout_width="match_parent"
-            android:layout_height="623dp"
-            android:layout_below="@+id/homeButton"
-            android:layout_centerHorizontal="true" />
-    </RelativeLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
-
